requirements_render.txt
psycopg2-binary==2.9.10
fastapi==0.115.12
uvicorn==0.34.3
gunicorn==21.2.0
jinja2==3.1.6
httpx==0.28.1
pydantic-settings==2.9.1
sqlalchemy==2.0.41
alembic==1.16.1
asyncpg==0.30.0
email-validator==2.2.0
ujson==5.10.0